requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
Flask-Migrate==4.0.5
plotly==5.18.0
pandas>=2.2.0
python-dateutil>=2.8.2
gunicorn==21.2.0
ofxparse>=0.21
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
